requirements.txt
instabot==0.117.0
Pillow==7.1.2
requests==2.23.0
requests-toolbelt==0.9.1